deploy/requirements.txt
boto3==1.6.3
PyYAML==3.12
